services/web/project/requirements.txt
alembic==0.8.10
appdirs==1.4.0
blinker==1.3
click==6.7
Flask-Assets==0.12
Flask-Compress
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-RQ==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy
Flask-SSLify==0.1.5
Flask-WTF==0.11
gunicorn==
itsdangerous==0.24
Jinja2==2.9.5
jsmin==2.2.1
jsonpickle==0.9.2
psycopg2==2.7
raygun4py==3.1.4
redis==2.10.5
requests==2.9.1
rq==0.7.1
SQLAlchemy==1.1.5
webassets==0.12.1
Werkzeug==0.11.15
WTForms==2.1
Authlib>=0.9
python-dotenv
Flask-Mail
